refactor decoders into trait
